+++ Strata/Source/Building_StairsUp.cs
@@ -125,6 +125,7 @@ namespace Strata
             StrataPortalUtility.TransferHaulDesignation(this, pawn);
             StrataPortalUtility.TryDeliverConstructionCargo(pawn);
             DraftedPortalPathing.NotifyPortalArrival(pawn);
+            PortalRelayChain.NotifyPortalArrival(pawn);
             MapComponent_RaidPursuit.NotifyPortalArrival(pawn, pawn.MapHeld);
         }